move validation functions into its own helper file
